Added 1 configuration variable and 2 new commands: (#45)
g:ghost_enable_sync (default: 1)

:GhostToggleSync
:GhostSync

By default, Ghost synchronizes the content of the Vim buffer with the browser's textarea (normal behavior). If ghost_enable_sync is set to 0, Ghost only updates the textarea when the :GhostSync command is issued. While editing, :GhostToggleSync may be used to disable / enable syncing.

The purpose of this modification is to allow the user to control exactly when the textarea should receive an update. This was done specifically for Discourse forums, in which other forum users can see someone is in the process of replying. Discourse does not allow users to disable this. Using vim-ghost with syncing disabled, nobody on a Discourse forum is made aware that a reply is being typed until the very last syncing of the reply's textarea.

Co-authored-by: Pierre Coupard <pierre.coupard@noptel.fi>